Feature security

+++ src/main/java/com/example/flowerstore/controllers/CartController.java
@@ -15,6 +15,7 @@ import com.example.flowerstore.dto.request.CartDTO;
 import com.example.flowerstore.dto.response.ApiResponse;
 import com.example.flowerstore.entites.Cart;
 import com.example.flowerstore.services.CartService;
+import com.example.flowerstore.security.SecurityUtils;
 
 import lombok.RequiredArgsConstructor;
 
@@ -25,33 +26,38 @@ public class CartController {
     private final CartService cartService;
 
     //add product to cart
-    @PostMapping("/carts/{userId}")
+    @PostMapping("/users/{userId}/carts")
     public ResponseEntity<ApiResponse<Cart>> createCart(@PathVariable Long userId, @RequestBody CartDTO cartDTO) {
+        SecurityUtils.validateUserAccess(userId);
         return ResponseEntity.ok(new ApiResponse<>(HttpStatus.OK.value(), "Cart created successfully", cartService.createCart(userId, cartDTO)));
     }
 
     //get cart by user id
-    @GetMapping("/carts/{userId}")
+    @GetMapping("/users/{userId}/carts")
     public ResponseEntity<ApiResponse<Cart>> getCartByUserId(@PathVariable Long userId) {
+        SecurityUtils.validateUserAccess(userId);
         return ResponseEntity.ok(new ApiResponse<>(HttpStatus.OK.value(), "Cart retrieved successfully", cartService.getCartByUserId(userId)));
     }
 
     //update cart
-    @PutMapping("/carts/{userId}")
+    @PutMapping("/users/{userId}/carts")
     public ResponseEntity<ApiResponse<Cart>> updateCart(@PathVariable Long userId, @RequestBody CartDTO cartDTO) {
+        SecurityUtils.validateUserAccess(userId);
         return ResponseEntity.ok(new ApiResponse<>(HttpStatus.OK.value(), "Cart updated successfully", cartService.updateCart(userId, cartDTO)));
     }
 
     //delete cart
-    @DeleteMapping("/carts/{userId}")
+    @DeleteMapping("/users/{userId}/carts")
     public ResponseEntity<ApiResponse<Void>> deleteCart(@PathVariable Long userId) {
+        SecurityUtils.validateUserAccess(userId);
         cartService.deleteCart(userId);
         return ResponseEntity.ok(new ApiResponse<>(HttpStatus.OK.value(), "Cart deleted successfully", null));
     }
 
     //delete product from cart
-    @DeleteMapping("/carts/{userId}/{productId}")
+    @DeleteMapping("/users/{userId}/carts/product/{productId}")
     public ResponseEntity<ApiResponse<Void>> deleteProductFromCart(@PathVariable Long userId, @PathVariable Long productId) {
+        SecurityUtils.validateUserAccess(userId);
         cartService.deleteProductFromCart(userId, productId);
         return ResponseEntity.ok(new ApiResponse<>(HttpStatus.OK.value(), "Product deleted from cart successfully", null));
     }
